+++ routes/bill.route.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const controller = require("../controllers/bill.controller");
+const controller = require("../controllers/control/bill.controller");
 
 router.get("/", controller.getAll);
 router.post("/", controller.add);
